+++ src/packages/piral-core/src/components/ExtensionSlot.tsx
@@ -2,10 +2,15 @@ import * as React from 'react';
 import { isfunc } from 'piral-base';
 import { useGlobalState } from '../hooks';
 import { defaultRender } from '../utils';
-import { ExtensionSlotProps } from '../types';
+import { ExtensionSlotProps, PiralExtensionSlotMap } from '../types';
 
-export function ExtensionSlot<T = any>({ name, render = defaultRender, empty, params }: ExtensionSlotProps<T>) {
-  const extensions = useGlobalState(s => s.registry.extensions[name] || []);
+export function ExtensionSlot<T extends keyof PiralExtensionSlotMap>({
+  name,
+  render = defaultRender,
+  empty,
+  params,
+}: ExtensionSlotProps) {
+  const extensions = useGlobalState(s => s.registry.extensions[name as string] || []);
   return render(
     extensions.length === 0 && isfunc(empty)
       ? [defaultRender(empty(), 'empty')]
@@ -20,4 +25,5 @@ export function ExtensionSlot<T = any>({ name, render = defaultRender, empty, pa
         )),
   );
 }
-(ExtensionSlot as React.FC<ExtensionSlotProps>).displayName = `ExtensionSlot`;
+
+ExtensionSlot.displayName = `ExtensionSlot`;