+++ src/packages/piral-core/src/modules/api.ts
@@ -27,7 +27,7 @@ export function createCoreApi(context: GlobalStateContext): ApiExtender<PiletCor
         context.unregisterPage(route);
       },
       registerExtension(name, arg, defaults) {
-        context.registerExtension(name, {
+        context.registerExtension(name as string, {
           pilet,
           component: withApi(context.converters, arg, api, 'extension'),
           reference: arg,
@@ -35,7 +35,7 @@ export function createCoreApi(context: GlobalStateContext): ApiExtender<PiletCor
         });
       },
       unregisterExtension(name, arg) {
-        context.unregisterExtension(name, arg);
+        context.unregisterExtension(name as string, arg);
       },
       renderHtmlExtension(element, props) {
         renderInDom(context, element, ExtensionSlot, props);