+++ src/packages/piral-core/src/types/api.ts
@@ -1,10 +1,9 @@
 import { ComponentType } from 'react';
 import { RouteComponentProps } from 'react-router-dom';
 import { PiletApi, Pilet, PiletMetadata, EventEmitter } from 'piral-base';
-import { Dict } from './common';
 import { PiletCustomApi } from './custom';
 import { AnyComponent } from './components';
-import { ExtensionSlotProps } from './extension';
+import { ExtensionSlotProps, PiralExtensionSlotMap } from './extension';
 import { SharedData, DataStoreOptions } from './data';
 
 export { PiletApi, Pilet, PiletMetadata, EventEmitter };
@@ -16,11 +15,11 @@ export interface BaseComponentProps {
   piral: PiletApi;
 }
 
-export interface ExtensionComponentProps<T = Dict<any>> extends BaseComponentProps {
+export interface ExtensionComponentProps<T extends keyof PiralExtensionSlotMap> extends BaseComponentProps {
   /**
    * The provided parameters for showing the extension.
    */
-  params: T;
+  params: PiralExtensionSlotMap[T];
 }
 
 export interface RouteBaseProps<UrlParams = any, UrlState = any>
@@ -67,14 +66,21 @@ export interface PiletCoreApi {
    * @param Component The component to be rendered.
    * @param defaults Optionally, sets the default values for the expected data.
    */
-  registerExtension<T>(name: string, Component: AnyComponent<ExtensionComponentProps<T>>, defaults?: T): void;
+  registerExtension<T extends keyof PiralExtensionSlotMap>(
+    name: T,
+    Component: AnyComponent<ExtensionComponentProps<T>>,
+    defaults?: T,
+  ): void;
   /**
    * Unregisters a global extension component.
    * Only previously registered extension components can be unregistered.
    * @param name The name of the extension slot to unregister from.
    * @param Component The registered extension component to unregister.
    */
-  unregisterExtension<T>(name: string, Component: AnyComponent<ExtensionComponentProps<T>>): void;
+  unregisterExtension<T extends keyof PiralExtensionSlotMap>(
+    name: T,
+    Component: AnyComponent<ExtensionComponentProps<T>>,
+  ): void;
   /**
    * React component for displaying extensions for a given name.
    */
@@ -84,7 +90,7 @@ export interface PiletCoreApi {
    * @param element The DOM element or shadow root as a container for rendering the extension.
    * @param props The extension's rendering props.
    */
-  renderHtmlExtension<T = any>(element: HTMLElement | ShadowRoot, props: ExtensionSlotProps<T>): void;
+  renderHtmlExtension(element: HTMLElement | ShadowRoot, props: ExtensionSlotProps): void;
 }
 
 declare module 'piral-base/lib/types' {