+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed robustness of the declaration generation
 * Improved typings for declaring extensions (using `PiralCustomExtensionSlotMap`)
 * Fixed Parcel logger verbose mode switch on Windows (#199)
+* Documented the use of `regenerator-runtime` (#200)
 * Added optional meta data argument for `registerPage` (#201)
 
 ## 0.11.2 (April 19, 2020)