+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed robustness of the declaration generation
 * Improved typings for declaring extensions (using `PiralCustomExtensionSlotMap`) (#197)
+* Added generic type for `PiralStoreDataEvent` (#198)
 * Fixed Parcel logger verbose mode switch on Windows (#199)
 * Documented the use of `regenerator-runtime` (#200)
 * Added optional meta data argument for `registerPage` (#201)